Separate static functions to utilities
